ProjectSourceCode/requirements.txt
requests==2.33.1
python-dotenv==1.2.2
finnhub-python==2.4.27
